refactor(actor): split actor cell facet modules
