+++ modules/actor-core-kernel/src/actor/actor_cell_dispatch_test.rs
@@ -1,0 +1,197 @@
+use super::*;
+
+#[test]
+fn identify_replies_with_actor_identity_without_invoking_actor() {
+  let system = ActorSystem::new_empty().state();
+  let actor_received = ArcShared::new(SpinSyncMutex::new(0usize));
+  let actor_replies = ArcShared::new(SpinSyncMutex::new(Vec::new()));
+  let actor_props = Props::from_fn({
+    let actor_received = actor_received.clone();
+    let actor_replies = actor_replies.clone();
+    move || IdentityProbeActor::new(actor_received.clone(), actor_replies.clone())
+  });
+  let target =
+    ActorCell::create(system.clone(), Pid::new(60, 0), None, "target".to_string(), &actor_props).expect("target");
+  let reply_received = ArcShared::new(SpinSyncMutex::new(0usize));
+  let reply_replies = ArcShared::new(SpinSyncMutex::new(Vec::new()));
+  let reply_props = Props::from_fn({
+    let reply_received = reply_received.clone();
+    let reply_replies = reply_replies.clone();
+    move || IdentityProbeActor::new(reply_received.clone(), reply_replies.clone())
+  });
+  let reply_to =
+    ActorCell::create(system.clone(), Pid::new(61, 0), None, "reply".to_string(), &reply_props).expect("reply");
+  system.register_cell(target.clone());
+  system.register_cell(reply_to.clone());
+
+  let mut invoker = ActorCellInvoker { cell: target.downgrade() };
+  let identify = Identify::new(AnyMessage::new("corr"));
+  let message = AnyMessage::new(identify).with_sender(reply_to.actor_ref());
+
+  invoker.invoke(message).expect("identify");
+
+  assert_eq!(*actor_received.lock(), 0, "identify should not reach the actor receive method");
+  wait_until(|| reply_replies.lock().len() == 1);
+  let replies = reply_replies.lock();
+  assert_eq!(replies.len(), 1);
+  let correlation_id = replies[0].correlation_id().payload().downcast_ref::<&str>().expect("&str");
+  assert_eq!(*correlation_id, "corr");
+  assert_eq!(replies[0].actor_ref().expect("actor ref").pid(), target.pid());
+}
+
+#[test]
+fn poison_pill_system_message_invokes_post_stop() {
+  let state = ActorSystem::new_empty().state();
+  let log = ArcShared::new(SpinSyncMutex::new(Vec::new()));
+  let props = Props::from_fn({
+    let log = log.clone();
+    move || LifecycleRecorderActor::new(log.clone())
+  });
+  let cell =
+    ActorCell::create(state.clone(), Pid::new(410, 0), None, "probe".to_string(), &props).expect("create actor cell");
+  state.register_cell(cell.clone());
+
+  let mut invoker = ActorCellInvoker { cell: cell.downgrade() };
+  invoker.system_invoke(SystemMessage::Create).expect("create");
+  invoker.system_invoke(SystemMessage::PoisonPill).expect("poison pill");
+
+  let snapshot = log.lock().clone();
+  assert_eq!(snapshot, vec!["pre_start", "post_stop"]);
+}
+
+#[test]
+fn poison_pill_public_message_invokes_post_stop() {
+  // Given: 起動済み actor に public PoisonPill payload を直接送る
+  let state = ActorSystem::new_empty().state();
+  let log = ArcShared::new(SpinSyncMutex::new(Vec::new()));
+  let props = Props::from_fn({
+    let log = log.clone();
+    move || LifecycleRecorderActor::new(log.clone())
+  });
+  let cell =
+    ActorCell::create(state.clone(), Pid::new(420, 0), None, "probe".to_string(), &props).expect("create actor cell");
+  state.register_cell(cell.clone());
+
+  let mut invoker = ActorCellInvoker { cell: cell.downgrade() };
+  invoker.system_invoke(SystemMessage::Create).expect("create");
+
+  // When: public message を通常 user message 経路で配送する
+  invoker.invoke(AnyMessage::new(PoisonPill)).expect("poison pill");
+
+  // Then: SystemMessage alias ではなくても auto-receive として停止処理が走る
+  let snapshot = log.lock().clone();
+  assert_eq!(snapshot, vec!["pre_start", "post_stop"]);
+}
+
+#[test]
+fn kill_system_message_reports_fatal_failure() {
+  let state = ActorSystem::new_empty().state();
+  let log = ArcShared::new(SpinSyncMutex::new(Vec::new()));
+  let props = Props::from_fn({
+    let log = log.clone();
+    move || LifecycleRecorderActor::new(log.clone())
+  });
+  let cell =
+    ActorCell::create(state.clone(), Pid::new(411, 0), None, "probe".to_string(), &props).expect("create actor cell");
+  state.register_cell(cell.clone());
+
+  let mut invoker = ActorCellInvoker { cell: cell.downgrade() };
+  invoker.system_invoke(SystemMessage::Create).expect("create");
+  let error = invoker.system_invoke(SystemMessage::Kill).expect_err("kill should report failure");
+
+  assert_eq!(error, ActorError::fatal("Kill"));
+}
+
+#[test]
+fn kill_public_message_reports_fatal_failure() {
+  // Given: 起動済み actor に public Kill payload を直接送る
+  let state = ActorSystem::new_empty().state();
+  let log = ArcShared::new(SpinSyncMutex::new(Vec::new()));
+  let props = Props::from_fn({
+    let log = log.clone();
+    move || LifecycleRecorderActor::new(log.clone())
+  });
+  let cell =
+    ActorCell::create(state.clone(), Pid::new(421, 0), None, "probe".to_string(), &props).expect("create actor cell");
+  state.register_cell(cell.clone());
+
+  let mut invoker = ActorCellInvoker { cell: cell.downgrade() };
+  invoker.system_invoke(SystemMessage::Create).expect("create");
+
+  // When: public message を通常 user message 経路で配送する
+  let error = invoker.invoke(AnyMessage::new(Kill)).expect_err("kill should fail");
+
+  // Then: runtime は public payload を fatal kill として扱う
+  assert_eq!(error, ActorError::fatal("Kill"));
+}
+
+#[test]
+fn poison_pill_user_message_preserves_user_ordering() {
+  let state = ActorSystem::new_empty().state();
+  let log = ArcShared::new(SpinSyncMutex::new(Vec::new()));
+  let props = Props::from_fn({
+    let log = log.clone();
+    move || LifecycleRecorderActor::new(log.clone())
+  });
+  let cell =
+    ActorCell::create(state.clone(), Pid::new(412, 0), None, "probe".to_string(), &props).expect("create actor cell");
+  state.register_cell(cell.clone());
+
+  let mut invoker = ActorCellInvoker { cell: cell.downgrade() };
+  invoker.system_invoke(SystemMessage::Create).expect("create");
+
+  assert!(cell.actor_ref().try_tell(AnyMessage::new(1_u8)).is_ok());
+  cell.actor_ref().poison_pill();
+  // Message 2 races against the dispatcher processing PoisonPill. It is
+  // either accepted (and later drained at close) or rejected with
+  // `SendError::Closed` if the mailbox has already been closed. Either way,
+  // the ordering invariant below requires that it is not received.
+  let second_result = cell.actor_ref().try_tell(AnyMessage::new(2_u8));
+  assert!(
+    second_result.is_ok() || matches!(second_result, Err(crate::actor::error::SendError::Closed(_))),
+    "message 2 should be accepted or rejected as Closed, got {second_result:?}",
+  );
+
+  wait_until(|| log.lock().len() >= 3);
+  let snapshot = log.lock().clone();
+  assert_eq!(snapshot, vec!["pre_start", "receive", "post_stop"]);
+}
+
+#[test]
+fn kill_user_message_reports_fatal_failure() {
+  let state = ActorSystem::new_empty().state();
+  let log = ArcShared::new(SpinSyncMutex::new(Vec::new()));
+  let props = Props::from_fn({
+    let log = log.clone();
+    move || LifecycleRecorderActor::new(log.clone())
+  });
+  let cell =
+    ActorCell::create(state.clone(), Pid::new(413, 0), None, "probe".to_string(), &props).expect("create actor cell");
+  state.register_cell(cell.clone());
+
+  let mut invoker = ActorCellInvoker { cell: cell.downgrade() };
+  invoker.system_invoke(SystemMessage::Create).expect("create");
+  let error = invoker.invoke(AnyMessage::new(SystemMessage::Kill)).expect_err("kill should fail");
+  assert_eq!(error, ActorError::fatal("Kill"));
+}
+
+#[test]
+fn system_queue_is_drained_before_user_queue() {
+  let state = ActorSystem::new_empty().state();
+  let log = ArcShared::new(SpinSyncMutex::new(Vec::new()));
+  let props = Props::from_fn({
+    let log = log.clone();
+    move || LifecycleRecorderActor::new(log.clone())
+  });
+  let cell =
+    ActorCell::create(state.clone(), Pid::new(42, 0), None, "probe".to_string(), &props).expect("create actor cell");
+  state.register_cell(cell.clone());
+
+  cell.new_dispatcher_shared().system_dispatch(&cell, SystemMessage::Create).expect("system enqueue");
+  assert!(cell.actor_ref().try_tell(AnyMessage::new(())).is_ok());
+
+  let _scheduled = cell.new_dispatcher_shared().register_for_execution(&cell.mailbox(), true, true);
+
+  let snapshot = log.lock().clone();
+  assert_eq!(snapshot, vec!["pre_start", "receive"]);
+}
